Using new nuntius define_handler
