+++ app/models/laboratory/event.rb
@@ -1,0 +1,5 @@
+module Laboratory
+  module Event
+    extend Mumukit::Nuntius::EventConsumer::Handler
+  end
+end
