--- app/models/laboratory/event/organization_created.rb
@@ -1,9 +1,0 @@
-module Laboratory
-  module Event
-    class OrganizationCreated
-      def self.execute!(payload)
-        Organization.create_from_json! payload.deep_symbolize_keys[:organization]
-      end
-    end
-  end
-end
