+++ app/models/laboratory/event/user.rb
@@ -1,0 +1,5 @@
+module Laboratory::Event
+  define_handler :UserChanged do |payload|
+    User.import_from_json! payload.deep_symbolize_keys[:user]
+  end
+end
