--- app/models/laboratory/event/user_changed.rb
@@ -1,9 +1,0 @@
-module Laboratory
-  module Event
-    class UserChanged
-      def self.execute!(payload)
-        User.import_from_json! payload.deep_symbolize_keys[:user]
-      end
-    end
-  end
-end
